Extract contract-file-at-ref helpers into a dedicated file
The architecture test now applies the 500-line production file cap to
every module, which surfaced protocol_minimum_version_guard.go at 523
lines. The git contract reading layer (legacy-path fallback and command
execution) is a distinct concern also consumed by the dispatcher guard,
so move it to its own file instead of raising the cap.
